eliminate some type conversion warnings
* oct-rl-edit.h, oct-rl-edit.c (octave_rl_add_defun): Change prototype
to match rl_add_defun.
(octave_rl_ctrl, octave_rl_meta): Return int instead of char.
* oct-rl-hist.c (octave_history_list): Use size_t as needed.

+++ liboctave/util/oct-rl-hist.c
@@ -279,13 +279,13 @@ octave_history_list (int limit, int number_lines)
 
       beg = (limit < 0 || end < limit) ? 0 : (end - limit);
 
-      retval = malloc ((end - beg + 1) * sizeof (char **));
+      retval = malloc ((size_t) (end - beg + 1) * sizeof (char **));
 
       k = 0;
       for (i = beg; i < end; i++)
         {
           char *line = hlist[i]->line;
-          int len = line ? strlen (line) : 0;
+          size_t len = line ? strlen (line) : 0;
           char *tmp = malloc (len + 64);
 
           if (number_lines)
